extracting the MockCatalog logic to use it everywhere
