+++ core/src/test/java/org/polypheny/db/docker/DockerInstanceTest.java
@@ -41,7 +41,7 @@ public class DockerInstanceTest {
     public static void initClass() {
         if ( Catalog.INSTANCE == null ) {
             // some functionality needs to use the catalog, so we use a mock
-            Catalog.setAndGetInstance( new MockCatalog() );
+            Catalog.setAndGetInstance( new MockCatalogDocker() );
         }
     }
